Rename task and queue
