+++ features/corpse_util.lua
@@ -1,6 +1,6 @@
 local Event = require 'utils.event'
 local Global = require 'utils.global'
-local Task = require 'utils.Task'
+local Task = require 'utils.threading'
 local Token = require 'utils.token'
 local Game = require 'utils.game'
