+++ map_gen/Diggy/Feature/AlienSpawner.lua
@@ -7,7 +7,7 @@ require 'utils.table'
 local Event = require 'utils.event'
 local Global = require 'utils.global'
 local Token = require 'utils.token'
-local Task = require 'utils.Task'
+local Task = require 'utils.threading'
 local AlienEvolutionProgress = require 'map_gen.Diggy.AlienEvolutionProgress'
 local Debug = require 'map_gen.Diggy.Debug'
 local Template = require 'map_gen.Diggy.Template'
