+++ map_gen/Diggy/Feature/DiggyCaveCollapse.lua
@@ -8,7 +8,7 @@ local Event = require 'utils.event'
 local Template = require 'map_gen.Diggy.Template'
 local ScoreTable = require 'map_gen.Diggy.ScoreTable'
 local Debug = require 'map_gen.Diggy.Debug'
-local Task = require 'utils.Task'
+local Task = require 'utils.threading'
 local Token = require 'utils.token'
 local Global = require 'utils.global'
 local Game = require 'utils.game'
