+++ map_gen/combined/dagobah_swamp.lua
@@ -3,7 +3,7 @@
 -- !! ATTENTION !!
 -- Use water only in starting area as map setting!!!
 local perlin = require 'map_gen.shared.perlin_noise'
-local Task = require 'utils.Task'
+local Task = require 'utils.threading'
 wreck_item_pool = {}
 wreck_item_pool = {
     {name = 'iron-gear-wheel', count = 32},
