+++ map_gen/misc/infinite_storage_chest.lua
@@ -1,7 +1,7 @@
 local Event = require 'utils.event'
 local Token = require 'utils.token'
 local Gui = require 'utils.gui'
-local Task = require 'utils.Task'
+local Task = require 'utils.threading'
 local Global = require 'utils.global'
 local Game = require 'utils.game'
