+++ map_gen/misc/terraforming.lua
@@ -1,6 +1,6 @@
 local Game = require 'utils.game'
 local Event = require 'utils.event'
-local Task = require 'utils.Task'
+local Task = require 'utils.threading'
 local Token = require 'utils.token'
 local random = math.random
 local insert = table.insert
