+++ map_gen/presets/crash_site/entity_died_events.lua
@@ -1,5 +1,5 @@
 local Event = require 'utils.event'
-local Task = require 'utils.Task'
+local Task = require 'utils.threading'
 local Token = require 'utils.token'
 local Global = require 'utils.global'
